+++ changes/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 03-05-23
+
+- integrated healthcheck into swagger/logic/service
+- exposed kafka configuration via environmental variable
+
 ## [1.0.3] - 02-22-23
 
 - fixed security vulnerabilities by updating modules